server/requirements.txt
openenv-core[core]>=0.2.3
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
PyTDC>=0.4.1
selfies>=2.1.1
rdkit>=2023.9.5
numpy>=1.24
